requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
translate==3.6.1
google-cloud-translate==3.24.0
